requirements.txt
# --- Django and REST Framework ---
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
sqlparse==0.5.3
tzdata==2025.2

# --- Production & Deployment ---
gunicorn==23.0.0
whitenoise==6.9.0
packaging==25.0

# --- Image and Face Recognition ---
face-recognition==1.3.0
dlib==19.24.6
cmake==3.30.3
opencv-python==4.10.0.84
numpy==2.1.3
pillow==11.2.1

# --- API Utilities ---
requests==2.32.3

# --- Optional (for QR code, Firebase integration, logging) ---
qrcode==8.2
firebase-admin==6.5.0
loguru==0.7.2
